requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]
python-multipart

# Database & ORM
sqlmodel
aiosqlite

# AI & External APIs
openai
groq
google-genai
python-dotenv

# Asynchronous File Handling
aiofiles